backend/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
ccxt>=4.0.0
pymongo[srv]>=4.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
pydantic>=2.0.0
certifi>=2024.0.0
